requirements.txt
opencv-python==4.10.0.84
python-dotenv==1.0.0
moviepy==1.0.3
openai==1.59.6
streamlit==1.38.0
yt_dlp==2025.1.15
authlib>=1.3.2